+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-05-05 20:18:20
-// Commit: feat(perf): integrate custom telemetry logger in perf
+// Updated at: 2026-05-05 21:41:07
+// Commit: style(perf): harmonize light and dark theme palette for perf
 
 class perfComponentState {
-  final String commitRef = '906';
-  final String lastMessage = 'feat(perf): integrate custom telemetry logger in perf';
-  final DateTime timestamp = DateTime.parse('2026-05-05T20:18:20+05:00');
+  final String commitRef = '911';
+  final String lastMessage = 'style(perf): harmonize light and dark theme palette for perf';
+  final DateTime timestamp = DateTime.parse('2026-05-05T21:41:07+05:00');
 }